+++ bindings/python/IK/src/SE3Task.cpp
@@ -16,6 +16,7 @@
 #include <BipedalLocomotion/System/ITaskControllerManager.h>
 
 #include <BipedalLocomotion/bindings/IK/SE3Task.h>
+#include <BipedalLocomotion/bindings/System/LinearTask.h>
 
 #include <manif/manif.h>
 
@@ -36,7 +37,9 @@ void CreateSE3Task(pybind11::module& module)
                IKLinearTask,
                BipedalLocomotion::System::ITaskControllerManager>(module, "SE3Task")
         .def(py::init())
-        .def("set_kin_dyn", &SE3Task::setKinDyn, py::arg("kin_dyn"))
+        .def("set_kin_dyn",
+             BipedalLocomotion::bindings::System::setKinDyn<SE3Task>,
+             py::arg("kin_dyn"))
         .def("set_set_point",
              &SE3Task::setSetPoint,
              py::arg("I_H_F"),