+++ bindings/python/IK/src/SO3Task.cpp
@@ -11,7 +11,9 @@
 
 #include <BipedalLocomotion/IK/IKLinearTask.h>
 #include <BipedalLocomotion/IK/SO3Task.h>
+
 #include <BipedalLocomotion/bindings/IK/SO3Task.h>
+#include <BipedalLocomotion/bindings/System/LinearTask.h>
 
 #include <manif/manif.h>
 
@@ -29,7 +31,9 @@ void CreateSO3Task(pybind11::module& module)
 
     py::class_<SO3Task, std::shared_ptr<SO3Task>, IKLinearTask>(module, "SO3Task")
         .def(py::init())
-        .def("set_kin_dyn", &SO3Task::setKinDyn, py::arg("kin_dyn"))
+        .def("set_kin_dyn",
+             BipedalLocomotion::bindings::System::setKinDyn<SO3Task>,
+             py::arg("kin_dyn"))
         .def("set_set_point",
              &SO3Task::setSetPoint,
              py::arg("I_R_F"),