+++ bindings/python/TSID/src/JointDynamicsTask.cpp
@@ -5,13 +5,17 @@
  * distributed under the terms of the BSD-3-Clause license.
  */
 
+#include <iDynTree/Model/JointState.h>
 #include <pybind11/eigen.h>
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 
-#include <BipedalLocomotion/TSID/TSIDLinearTask.h>
 #include <BipedalLocomotion/TSID/JointDynamicsTask.h>
+#include <BipedalLocomotion/TSID/TSIDLinearTask.h>
+
+#include <BipedalLocomotion/bindings/System/LinearTask.h>
 #include <BipedalLocomotion/bindings/TSID/JointDynamicsTask.h>
+#include <BipedalLocomotion/bindings/TSID/TSIDLinearTask.h>
 
 namespace BipedalLocomotion
 {
@@ -29,7 +33,9 @@ void CreateJointDynamicsTask(pybind11::module& module)
                std::shared_ptr<JointDynamicsTask>, //
                TSIDLinearTask>(module, "JointDynamicsTask")
         .def(py::init())
-        .def("set_kin_dyn", &JointDynamicsTask::setKinDyn, py::arg("kin_dyn"));
+        .def("set_kin_dyn",
+             BipedalLocomotion::bindings::System::setKinDyn<JointDynamicsTask>,
+             py::arg("kin_dyn"));
 }
 
 } // namespace TSID